+++ src/components/navigation-accordion/__tests__/navigation-accordion-test-cases.js
@@ -35,11 +35,9 @@ testCases.withTags = {
             customTagProps: {
               customLabel: 'Custom',
               customTooltipText: 'This is a custom tag.',
-              customStyles: {
-                background: '#FEDADA',
-                color: '#bb2224',
-                borderColor: '#FD8383'
-              },
+              customBackground: 'bg-red-light',
+              customColor: 'color-red-dark',
+              customBorder: 'border--red',
               customIcon: 'alert'
             }
           }